Modular Systems Implemented
Finished Saving Systems to work with player files. This allows the locations to be interactive with the player PER SAVE. Which allows for construction and banning from places (Banning not yet implemented).

I also redid all the location code so now you can just make a new location in the "locationsSettings.json" file.